compatibility

+++ external/mt19937ar.h
@@ -4,18 +4,18 @@
 
 struct mersennetwister
 {
-    unsigned long mt[MERSENNETWISTER_N]; /* the array for the state vector  */
+    uint32_t mt[MERSENNETWISTER_N]; /* the array for the state vector  */
     int mti;    /* =MERSENNETWISTER_N+1; * mti==N+1 means mt[N] is not initialized */
 };
 
 
-void mersennetwister_init_genrand(struct mersennetwister *mt, unsigned long s);
-void mersennetwister_init_by_array(struct mersennetwister *mt, unsigned long init_key[], int key_length);
-unsigned long mersennetwister_genrand_int32(struct mersennetwister *mt);
-unsigned long mersennetwister_genrand_int32_strong(struct mersennetwister *mt, unsigned long mx);
+void mersennetwister_init_genrand(struct mersennetwister *mt, uint32_t s);
+void mersennetwister_init_by_array(struct mersennetwister *mt, uint32_t init_key[], int key_length);
+uint32_t mersennetwister_genrand_int32(struct mersennetwister *mt);
+uint32_t mersennetwister_genrand_int32_strong(struct mersennetwister *mt, uint32_t mx);
 long mersennetwister_genrand_int31(struct mersennetwister *mt);
 double mersennetwister_genrand_real1(struct mersennetwister *mt);
 double mersennetwister_genrand_real2(struct mersennetwister *mt);
 double mersennetwister_genrand_real3(struct mersennetwister *mt);
 double mersennetwister_genrand_res53(struct mersennetwister *mt);
-unsigned long long mersennetwister_genrand_int64(struct mersennetwister *mt);
+uint64_t mersennetwister_genrand_int64(struct mersennetwister *mt);
